+++ internal/interceptor/inbound_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package transportinterceptor
+package interceptor
 
 import (
 	"context"
@@ -28,22 +28,12 @@ import (
 	"go.uber.org/yarpc/api/transport"
 )
 
+// UnaryHandlerFunc allows a function to be treated as a UnaryHandler for testing purposes.
 type UnaryHandlerFunc func(ctx context.Context, req *transport.Request, resw transport.ResponseWriter) error
 
+// Handle calls the underlying function in UnaryHandlerFunc.
 func (f UnaryHandlerFunc) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter) error {
 	return f(ctx, req, resw)
-}
-
-type OnewayHandlerFunc func(ctx context.Context, req *transport.Request) error
-
-func (f OnewayHandlerFunc) HandleOneway(ctx context.Context, req *transport.Request) error {
-	return f(ctx, req)
-}
-
-type StreamHandlerFunc func(s *transport.ServerStream) error
-
-func (f StreamHandlerFunc) HandleStream(s *transport.ServerStream) error {
-	return f(s)
 }
 
 // TestNopUnaryInbound ensures NopUnaryInbound calls the underlying handler without modification.
@@ -74,70 +64,6 @@ func TestApplyUnaryInbound(t *testing.T) {
 
 	wrappedHandler := ApplyUnaryInbound(handler, middleware)
 	err := wrappedHandler.Handle(context.Background(), &transport.Request{}, nil)
-	assert.NoError(t, err)
-	assert.True(t, called)
-}
-
-// TestNopOnewayInbound ensures NopOnewayInbound calls the underlying handler without modification.
-func TestNopOnewayInbound(t *testing.T) {
-	var called bool
-	handler := OnewayHandlerFunc(func(ctx context.Context, req *transport.Request) error {
-		called = true
-		return nil
-	})
-
-	err := NopOnewayInbound.HandleOneway(context.Background(), &transport.Request{}, handler)
-	assert.NoError(t, err)
-	assert.True(t, called)
-}
-
-// TestApplyOnewayInbound ensures that OnewayInbound middleware wraps correctly.
-func TestApplyOnewayInbound(t *testing.T) {
-	var called bool
-	handler := OnewayHandlerFunc(func(ctx context.Context, req *transport.Request) error {
-		called = true
-		return nil
-	})
-
-	middleware := OnewayInboundFunc(func(ctx context.Context, req *transport.Request, h transport.OnewayHandler) error {
-		assert.False(t, called)
-		return h.HandleOneway(ctx, req)
-	})
-
-	wrappedHandler := ApplyOnewayInbound(handler, middleware)
-	err := wrappedHandler.HandleOneway(context.Background(), &transport.Request{})
-	assert.NoError(t, err)
-	assert.True(t, called)
-}
-
-// TestNopStreamInbound ensures NopStreamInbound calls the underlying handler without modification.
-func TestNopStreamInbound(t *testing.T) {
-	var called bool
-	handler := StreamHandlerFunc(func(s *transport.ServerStream) error {
-		called = true
-		return nil
-	})
-
-	err := NopStreamInbound.HandleStream(&transport.ServerStream{}, handler)
-	assert.NoError(t, err)
-	assert.True(t, called)
-}
-
-// TestApplyStreamInbound ensures that StreamInbound middleware wraps correctly.
-func TestApplyStreamInbound(t *testing.T) {
-	var called bool
-	handler := StreamHandlerFunc(func(s *transport.ServerStream) error {
-		called = true
-		return nil
-	})
-
-	middleware := StreamInboundFunc(func(s *transport.ServerStream, h transport.StreamHandler) error {
-		assert.False(t, called)
-		return h.HandleStream(s)
-	})
-
-	wrappedHandler := ApplyStreamInbound(handler, middleware)
-	err := wrappedHandler.HandleStream(&transport.ServerStream{})
 	assert.NoError(t, err)
 	assert.True(t, called)
 }
